Prometheus: option to ignore entities in STATE_UNKNOWN
Giving a value of 0 by default can lead to erroneous
data being exported. For example, if a MQTT temperature
sensor is in `STATE_UNKNOWN` (which can happend after a
HASS restart), a temperature of 0°C will be exported.
I think it's better to export no value than a wrong one.